requirements.txt
google-cloud-storage==2.10.0
psycopg2-binary==2.9.7
functions-framework==3.4.0
